cleanup, refactor
